Implement error<> template
This commit adds the error template to be used with error enums. It
comes with several small helpers, most of which are not part of the
public interface. The only exception is overloaded.hpp, which is not
required for the implementation, but provides a useful helper for
visitation. It is used in the unit tests.